Refactor collab svc backend
